+++ src/App/App.scss
@@ -17,3 +17,7 @@
     transform: rotate(360deg);
   }
 }
+
+.App {
+  background-color: #343a40;
+}
