--- src/Components/Pages/Background/Background.scss
@@ -1,26 +1,0 @@
-.desktopBackground {
-  min-height: 100vh;
-  background-image: url(https://github.com/phillipsja97/personal-site-react/blob/master/src/Assets/blackWorldLines.jpg?raw=true);
-  background-attachment: fixed;
- //  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
- }
-
- .tabletBackground {
-  min-height: 100vh;
-  background-image: url(https://github.com/phillipsja97/personal-site-react/blob/master/src/Assets/blackWorldLines.jpg?raw=true);
-  background-attachment: fixed;
- //  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  
- }
-
- .phoneBackground {
-  min-height: 100vh;
-  background-image: url(https://github.com/phillipsja97/personal-site-react/blob/master/src/Assets/blackWorldLines.jpg?raw=true);
-  background-attachment: fixed;
- //  background-position: center;
-  background-repeat: no-repeat;
- }
