+++ src/Components/Pages/Contact/Contact.js
@@ -1,22 +1,20 @@
 import React from 'react';
 import MediaQuery from 'react-responsive';
-import Fade from 'react-reveal/Fade';
+import FadeIn from 'react-reveal/Fade';
 import { Button } from 'react-bootstrap';
 import { IconContext } from 'react-icons';
 import { FaGithub, FaLinkedin } from 'react-icons/fa';
-import Background from '../Background/Background';
 import './Contact.scss';
 
 class Contact extends React.Component {
   render() {
     return (
       <React.Fragment className="fluid-container">
-       <MediaQuery minDeviceWidth={1224}>
-          <Background />
-        <Fade bottom cascade>
         <div className="contactIcons">
           <h1 class="title">jamiephillips <span>/contact-me/</span><span2> (master):</span2></h1>
-           <div class="contactContainer">
+        <FadeIn>
+            <div className="box">
+              <div class="contactContainer">
                 <IconContext.Provider value={{ color: 'rgb(19, 228, 228)', className: 'githubContact' }}>
                   <div>
                     <a href='https://github.com/phillipsja97'><FaGithub /></a>
@@ -28,10 +26,10 @@ class Contact extends React.Component {
                     <a href='https://www.linkedin.com/in/jamieaphillips/'><FaLinkedin /></a>
                   </div>
                 </IconContext.Provider>
+              </div>
             </div>
+        </FadeIn>
         </div>
-        </Fade>
-        </MediaQuery>
     </React.Fragment>
     );
   }
