+++ src/Components/Pages/Projects/Projects.js
@@ -1,8 +1,8 @@
 /* eslint-disable max-len */
 import React from 'react';
 import MediaQuery from 'react-responsive';
-import Fade from 'react-reveal/Fade';
-import Background from '../Background/Background';
+import Fade from 'react-reveal';
+import Background from '../HomePage/HomePage';
 import projectData from '../../../Helpers/data/projectData';
 import ProjectCard from '../../Shared/ProjectCard/ProjectCard';
 import './Projects.scss';
@@ -12,35 +12,20 @@ class Projects extends React.Component {
     projects: [],
   }
 
-  getProjects = () => {
-    projectData.getProjects()
-      .then((projects) => {
-        this.setState({ projects });
-      })
-      .catch((errorFromProj) => console.error(errorFromProj));
-  }
-
-  componentDidMount() {
-    this.getProjects();
-  }
-
   render() {
+    const { projects } = this.props;
     return (
       <React.Fragment className="fluid-container">
-       <MediaQuery minDeviceWidth={1224}>
-          <Background />
-        <Fade bottom cascade>
         <div className="container-fluid" id="projSection">
-          <br></br>
             <p className="projTitle"> jamiephillips<span> ~/projects-I-built/</span><span2> (master):</span2></p>
-            <br></br>
+          <Fade bottom cascade>
           <div className="projectArea">
-           { this.state.projects.map((project) => <ProjectCard key={project.id} project={project} />)};
+          <div className="projContainer">
+           { projects.map((project) => <ProjectCard key={project.id} project={project} />)};
+           </div>
           </div>
-          <br></br><br></br>
+         </Fade>
         </div>
-        </Fade>
-        </MediaQuery>
       </React.Fragment>
     );
   }
