+++ src/Components/Pages/Technologies/Technologies.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import MediaQuery from 'react-responsive';
 import Fade from 'react-reveal/Fade';
-import Background from '../Background/Background';
+import Background from '../HomePage/HomePage';
 import TechnologiesCard from '../../Shared/TechnologiesCard/TechnologiesCard';
 import techData from '../../../Helpers/data/techData';
 import './Technologies.scss';
@@ -12,33 +12,21 @@ class Technologies extends React.Component {
     technologies: [],
   }
 
-  getTechnologies = () => {
-    techData.getTech()
-      .then((technologies) => {
-        this.setState({ technologies });
-      })
-      .catch((errorFromTech) => console.error(errorFromTech));
-  }
-
-  componentDidMount() {
-    this.getTechnologies();
-  }
-
   render() {
+    const { technologies } = this.props;
     return (
       <React.Fragment className="fluid-container">
         <div className="technologies">
-        <MediaQuery minDeviceWidth={1224}>
-          <Background />
-              <Fade bottom cascade>
                 <div className="images" id="tech">
                   <p className="techTitle"> jamiephillips<span> ~/technologies-I-have-used/</span><span2> (master):</span2></p>
-                  <div className="techArea">
-                      { this.state.technologies.map((tech) => <TechnologiesCard key={tech.id} tech={tech} />)};
-                  </div>
+                    <Fade bottom cascade>
+                      <div className="empty">
+                        <div className="techArea">
+                            { technologies.map((tech) => <TechnologiesCard key={tech.id} tech={tech} />)};
+                        </div>
+                      </div>
+                    </Fade>
                 </div>
-              </Fade>
-        </MediaQuery>
         </div>
       </React.Fragment>
     );
